VirtualPopulation R6 Class

The VirtualPopulation class provides functionality to:

Create virtual populations based on an inferred multivariate distribution and demographic constraints.

Add additive or multiplicative inter-occasion variability (IOV) to model parameters.

Simulate model outputs for a virtual population using reference and test simulation files.

Value

An object of class VirtualPopulation.

Details

A class for generating and simulating virtual populations with support for inter-occasion variability (IOV).

Public Methods

initialize(inferredDistribution, demographyRanges, numberOfVirtualIndividuals, proportionOfFemales, numberOfClusters): Constructor for the class.

addAdditiveIOV(parameterPath, SD): Add additive IOV to a specified parameter.

addMultiplicativeIOV(parameterPath, CV): Add multiplicative IOV to a specified parameter.

simulateVirtualPopulation(referenceSimulationFilePath, testSimulationFilePath, numberOfReplicates, outputPath, startTime, endTime, resolutionPtsMin): Simulate the virtual population using specified input files and time settings.

Active Bindings

inferredDistribution: Returns the distribution used to generate the virtual population.

referencePopulationDataframe: Returns the reference population data frame.

testPopulationDataframe: Returns the test population data frame.

iovListSD: Returns the list of additive IOV values.

iovListCV: Returns the list of multiplicative IOV values.

Active bindings

inferredDistribution

Get the distribution object used to generate the virtual population.

referencePopulationDataframe

Get the data frame of the reference virtual population.

testPopulationDataframe

Get the data frame of the test virtual population.

iovListSD

Get the list of additive IOV standard deviations by parameter.

iovListCV

Get the list of multiplicative IOV coefficients of variation by parameter.

Methods

Public methods

VirtualPopulation$new()

VirtualPopulation$addPopulationDataFrames()

VirtualPopulation$checkPopulationDataFrames()

VirtualPopulation$addAdditiveIOV()

VirtualPopulation$addMultiplicativeIOV()

VirtualPopulation$simulateVirtualPopulation()

VirtualPopulation$simulateVirtualPopulationFromPKML()

VirtualPopulation$clone()

Method new()

Initialize a new VirtualPopulation object.

Usage

VirtualPopulation$new(
  inferredDistribution = NULL,
  demographyRanges = NULL,
  numberOfVirtualIndividuals = NULL,
  proportionOfFemales = 50,
  numberOfClusters = 4
)

Arguments

inferredDistribution

A distribution object for generating virtual individuals.

demographyRanges

A ospsuiteR PopulationCharacteristics R6 object OR a list specifying demographic constraints (e.g., age range, weight limits).

numberOfVirtualIndividuals

Integer. Number of virtual individuals to simulate. Default is 100.

proportionOfFemales

Numeric. Proportion (%) of females in the virtual population. Default is 50.

numberOfClusters

Integer. Number of clusters to use for stratifying the distribution. Default is 4.

Returns

A new VirtualPopulation object.

Method addPopulationDataFrames()

Adds dataframes for the virtual population to be simulated. Both a reference and test population dataframe are input to alow for possibly different parameter paths in the corresponding reference and test simulations.

Usage

VirtualPopulation$addPopulationDataFrames(
  referencePopulationDataframe,
  testPopulationDataframe
)

Arguments

referencePopulationDataframe

A PK-Sim population dataframe, in which all parameter paths are found in the reference simulation.

testPopulationDataframe

A PK-Sim population dataframe, in which all parameter paths are found in the test simulation.

Returns

None. Adds referencePopulationDataframe and testPopulationDataframe as dataframes to be used when simulating the reference and test models.

Method checkPopulationDataFrames()

Check that population dataframes have been set up, either by supplying demographyRanges at initialization of the VirtualPopulation or by supplying population data frames via the function addPopulationDataFrames.

Usage

VirtualPopulation$checkPopulationDataFrames()

Returns

None.

Method addAdditiveIOV()

Add additive inter-occasion variability (IOV) to a specific parameter in the virtual population.

Usage

VirtualPopulation$addAdditiveIOV(parameterPath, SD)

Arguments

parameterPath

Character. The path to the parameter in the model to which additive IOV is applied.

SD

Numeric. Standard deviation of the additive IOV. Must be a single positive number.

Returns

None. Adds to internal IOV list.

Examples

\dontrun{
vpop$addAdditiveIOV("PATH|TO|PARAMETER1", SD = 0.1)
}

Method addMultiplicativeIOV()

Add multiplicative inter-occasion variability (IOV) to a specific parameter.

Usage

VirtualPopulation$addMultiplicativeIOV(parameterPath, CV)

Arguments

parameterPath

Character. The path to the parameter in the model.

CV

Numeric. Coefficient of variation (CV) for the multiplicative IOV. Must be a single positive number.

Returns

None. Adds to internal IOV list.

Examples

\dontrun{
vpop$addMultiplicativeIOV("PATH|TO|PARAMETER2", CV = 0.2)
}

Method simulateVirtualPopulation()

Run simulations on the virtual population using provided simulation templates.

Usage

VirtualPopulation$simulateVirtualPopulation(
  referenceSimulationFilePath,
  testSimulationFilePath,
  numberOfReplicates = 1,
  outputPath,
  startTime,
  endTime,
  resolutionPtsMin
)

Arguments

referenceSimulationFilePath

Character. File path to the reference model simulation template.

testSimulationFilePath

Character. File path to the test model simulation template.

numberOfReplicates

Integer. Number of replicate simulations per virtual subject. Default is 1.

outputPath

Character. Directory to store simulation results.

startTime

Numeric. Start time of the simulation window.

endTime

Numeric. End time of the simulation window.

resolutionPtsMin

Numeric. Output resolution in points per minute.

Returns

A list containing results from reference and test population simulations.

Examples

\dontrun{
vpop$simulateVirtualPopulation(
  referenceSimulationFilePath = "ref.sim",
  testSimulationFilePath = "test.sim",
  outputPath = "PATH|TO|OUTPUT",
  startTime = 0,
  endTime = 1440,
  resolutionPtsMin = 10
)
}

Method simulateVirtualPopulationFromPKML()

Run simulations on the virtual population using provided simulation templates.

Usage

VirtualPopulation$simulateVirtualPopulationFromPKML(
  referenceSimulation,
  testSimulation,
  numberOfReplicates = 1,
  outputPath,
  startTime,
  endTime,
  resolutionPtsMin
)

Arguments

referenceSimulation

Reference simulation object.

testSimulation

Test simulation object.

numberOfReplicates

Integer. Number of replicate simulations per virtual subject. Default is 1.

outputPath

Character. Directory to store simulation results.

startTime

Numeric. Start time of the simulation window.

endTime

Numeric. End time of the simulation window.

resolutionPtsMin

Numeric. Output resolution in points per minute.

Returns

A list containing results from reference and test population simulations.

Examples

\dontrun{
vpop$simulateVirtualPopulationFromPKML(
  referenceSimulation = "refSim",
  testSimulation = "testSim",
  outputPath = "PATH|TO|OUTPUT",
  startTime = 0,
  endTime = 1440,
  resolutionPtsMin = 10
)
}

Method clone()

The objects of this class are cloneable with this method.

Usage

VirtualPopulation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
vpop <- VirtualPopulation$new(
  inferredDistribution = inferredDistribution,
  demographyRanges = list(age = c(18, 65)),
  numberOfVirtualIndividuals = 200,
  proportionOfFemales = 60,
  numberOfClusters = 5
)
vpop$addAdditiveIOV("PATH|TO|PARAMETER1", SD = 0.1)
vpop$addMultiplicativeIOV("PATH|TO|PARAMETER2", CV = 0.2)
vpop$simulateVirtualPopulation(
  referenceSimulationFilePath = "path/to/ref.sim",
  testSimulationFilePath = "path/to/test.sim",
  outputPath = "sim_outputs",
  startTime = 0,
  endTime = 24,
  resolutionPtsMin = 10
)
} # }

## ------------------------------------------------
## Method `VirtualPopulation$addAdditiveIOV`
## ------------------------------------------------

if (FALSE) { # \dontrun{
vpop$addAdditiveIOV("PATH|TO|PARAMETER1", SD = 0.1)
} # }

## ------------------------------------------------
## Method `VirtualPopulation$addMultiplicativeIOV`
## ------------------------------------------------

if (FALSE) { # \dontrun{
vpop$addMultiplicativeIOV("PATH|TO|PARAMETER2", CV = 0.2)
} # }

## ------------------------------------------------
## Method `VirtualPopulation$simulateVirtualPopulation`
## ------------------------------------------------

if (FALSE) { # \dontrun{
vpop$simulateVirtualPopulation(
  referenceSimulationFilePath = "ref.sim",
  testSimulationFilePath = "test.sim",
  outputPath = "PATH|TO|OUTPUT",
  startTime = 0,
  endTime = 1440,
  resolutionPtsMin = 10
)
} # }

## ------------------------------------------------
## Method `VirtualPopulation$simulateVirtualPopulationFromPKML`
## ------------------------------------------------

if (FALSE) { # \dontrun{
vpop$simulateVirtualPopulationFromPKML(
  referenceSimulation = "refSim",
  testSimulation = "testSim",
  outputPath = "PATH|TO|OUTPUT",
  startTime = 0,
  endTime = 1440,
  resolutionPtsMin = 10
)
} # }